requirements.txt
adal
msrestaazure
azure.keyvault